+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All Notable changes to `league-uri-parser` will be documented in this file
 
-## Next - TBD
+## 1.4.0 - 2018-03-14
 
 ### Added